+++ TextChat/Runtime/Scripts/ITextChatSystem.cs
@@ -1,7 +1,5 @@
 using Reflectis.SDK.Core;
-using Reflectis.SDK.TextChat.Enums;
 using System;
-using System.Collections.Generic;
 
 namespace Reflectis.SDK.TextChat
 {
@@ -10,10 +8,9 @@ namespace Reflectis.SDK.TextChat
     /// </summary>
     public interface ITextChatSystem : ISystem
     {
-        string[] Channels { get; }
+        ChatRoom[] ActiveChatChannels { get; }
 
-        bool IsConnected { get; set; }
-        bool IgnoreNotifications { get; set; }
+        bool IsConnected { get; }
 
         #region Events
 
@@ -32,46 +29,21 @@ namespace Reflectis.SDK.TextChat
         #endregion
 
         #region Message events
-        /// <summary>
-        /// Event invoked when a message is sent.
-        /// It will pass the name of the person/channel who
-        /// send it and the content of the message
-        /// </summary>
-        event Action<string, string> OnTxtMsgSent;
 
         /// <summary>
         /// Event invoked when a message is received. 
-        /// It will pass the name of the person/channel who
-        /// sent it, the content of the message and for who it is
+        /// It will pass the message and the channel in which the message was sent
         /// </summary>
-        event Action<string, string, string> OnTxtMsgReceived;
+        event Action<ChatRoom> OnTxtMsgReceived;
 
-
-        /// <summary>
-        /// Event invoked when the history messages are fetched.
-        /// It will pass a list with all the ChatMessages sent.
-        /// </summary>
-        event Action<List<ChatMessage>> OnTxtMsgFetched;
         #endregion
 
         #region Channel events
 
         /// <summary>
-        /// Event invoked when all the public channels are fetched from the server.
-        /// It will pass a list containing all the chat rooms
-        /// </summary>
-        event Action<List<ChatRoom>> OnAllChannelsFetched;
-
-        /// <summary>
         /// Event invoked when a user joins a channel
         /// </summary>
-        event Action<string> OnJoinedChannel;
-
-        /// <summary>
-        /// Event invoked when a public channel is fetched from the server.
-        /// It will pass the channel
-        /// </summary>
-        event Action<ChatRoom> OnChannelFetched;
+        event Action<ChatRoom> OnPublicRoomJoin;
 
         #endregion
 
@@ -79,23 +51,25 @@ namespace Reflectis.SDK.TextChat
 
         #region Methods
         /// <summary>
-        /// Connect the current user to the text chat.
+        /// Connect to text chat 
         /// </summary>
-        /// <param name="userId">Id of the user to connect to the text chat</param>
-        /// <param name="token">A token use to authorize the connection if needed</param>
-        void ConnectUserToTextChat(string userId, string token = "");
+        void Connect();
 
         /// <summary>
-        /// Keep the connection alive so a user can get incoming messages continuously
+        /// Wheter the user can chat or not
         /// </summary>
-        void KeepConnectionAlive();
+        /// <returns></returns>
         bool CanChat();
-        bool CanChatInChannel(string channelId);
+        /// <summary>
+        /// Wheter the user can chat or not in the given channel
+        /// </summary>
+        /// <returns></returns>
+        bool CanChatInRoom(string channelId);
 
         /// <summary>
         /// Disconnect the current user to the text chat
         /// </summary>
-        void DisconnectCurrentUser();
+        void Disconnect();
 
         /// <summary>
         /// Sends a new message to a specific user
@@ -111,34 +85,6 @@ namespace Reflectis.SDK.TextChat
         /// <param name="msg">A message</param>
         void SendMessageToChannel(string channelId, ChatMessage msg);
 
-        /// <summary>
-        /// Get all the history messages sent through a specific channel
-        /// </summary>
-        /// <param name="channelId">Id of the channel to fetch all the messages</param>
-        /// <param name="type">The type of the channel</param>
-        void FetchHistoryMessages(string channelId, EChatMessageType type);
-
-        /// <summary>
-        /// Get the current messages count from a specific channel.
-        /// </summary>
-        /// <param name="channelId">Id of the channel to fetch all the messages</param>
-        /// <param name="type">The type of the channel</param>
-        int GetCurrentMessagesCount(string channelId, EChatMessageType type);
-
-        /// <summary>
-        /// Get all the public channels
-        /// </summary>
-        void FetchAllChannels();
-
-        /// <summary>
-        /// Get all the private channels ids
-        /// </summary>
-        List<string> GetPrivateChannelIds();
-
-        /// <summary>
-        /// Joins the default channels
-        /// </summary>
-        void JoinTextChannel();
 
         /// <summary>
         /// Joins a text channel
@@ -146,18 +92,17 @@ namespace Reflectis.SDK.TextChat
         /// <param name="channelId">Id of the channel</param>
         void JoinTextChannel(string channelId);
 
-
         /// <summary>
-        /// Get info from a specific channel
+        /// Get info from a specific room
         /// </summary>
         /// <param name="channelId"></param>
-        void FetchChannelInfo(string channelId);
+        public ChatRoom GetPublicRoom(string channelId);
 
         /// <summary>
         /// Leave a text channel
         /// </summary>
         /// <param name="channelId">Id of the channel</param>
-        void LeaveTextChannel(string channelId);
+        void LeavePublicRoom(string channelId);
 
         /// <summary>
         /// Delete all messages from a specific conversation.
@@ -165,7 +110,7 @@ namespace Reflectis.SDK.TextChat
         /// </summary>
         /// <param name="conversationId"></param>
         /// <param name="type"></param>
-        void DeleteConversationFromServer(string conversationId, EChatMessageType type);
+        //void DeleteConversationFromServer(string conversationId, EChatMessageType type);
 
         #endregion
     }